requirements.txt
selenium
pytest
appium-python-client